fix: common apertures, centroids
refactor: create iop4lib.utils.photometry
refactor: imports
